decompress: enable post decompress filters

+++ src/lib/jp2/t1/T1DecompressScheduler.cpp
@@ -35,7 +35,7 @@ bool T1DecompressScheduler::prepareScheduleDecompress(TileComponent *tilec, Tile
 		GRK_ERROR( "Not enough memory for tile data");
 		return false;
 	}
-	for (uint32_t resno = 0; resno < tilec->resolutions_to_decompress; ++resno) {
+	for (uint8_t resno = 0; resno < tilec->resolutions_to_decompress; ++resno) {
 		auto res = &tilec->resolutions[resno];
 		for (uint8_t bandIndex = 0; bandIndex < res->numBandWindows; ++bandIndex) {
 			Subband *GRK_RESTRICT band = res->bandWindow + bandIndex;
